requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.6.0
django-jazzmin==3.0.1
channels==4.2.0
daphne==4.1.2
asgiref==3.11.1
psycopg2-binary==2.9.10
dj-database-url==2.2.0
python-dotenv==1.1.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
whitenoise==6.7.0
reportlab==4.2.2
requests==2.32.3
pillow==11.1.0
PyJWT==2.12.0
redis==7.3.0
channels_redis==4.2.1
gunicorn==23.0.0
